executor: delete WeightCacheManager; replace with WeightCaches in executor.h
Phase 5 — final WeightCacheManager deletion. The type is replaced by a
WeightCaches struct defined directly in executor.h with identical field
layout. GemmContext now takes const WeightCaches* instead of
const WeightCacheManager*. The free() method is inlined into
free_buffers() in executor_workspace_buffers.cu. Mode flags
(use_fp8, nvfp4_decode_mode, dual_path_quant) remain on WeightCaches
and are mirrored into hints_ as before. No behavioral change.

- Removes weight_cache_manager.{h,cu} (193 lines)
- Removes IMP_GRAPH_SOURCES entry in CMakeLists.txt
- Clears PHASE-3-TODO stale markers in executor_workspace.cu,
  executor_attention.cu, executor_ffn.cu
- Updates test DualPathQuant::WeightCacheManagerFlag → WeightCachesFlag
- 580 tests pass, 23 skipped (model-file-dependent)
