+++ CMakeLists.txt
@@ -204,7 +204,6 @@ set(IMP_GRAPH_SOURCES
     src/graph/executor_ssm_gdn.cu
     src/graph/executor_forward_moe.cu
     src/graph/expert_cache.cu
-    src/graph/weight_cache_manager.cu
     src/graph/weight_handle.cu
     src/graph/moe_workspace.cu
     src/graph/quant_scratch.cu
